refactor: extract TaskItem and TaskDetailModal into separate classes
